Generate a json schema for the serialized analysis. Assert that the analyses we generate are legal under it.
