+++ build-tools/fixup-schema.js
@@ -1,0 +1,21 @@
+/**
+ * @license
+ * Copyright (c) 2015 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+'use strict';
+const fs = require('fs');
+const path = 'lib/analysis.schema.json';
+const schema = JSON.parse(fs.readFileSync(path, 'utf-8'));
+schema.definitions.Event.properties.detail.properties.properties.items.$ref =
+    '#/definitions/Property';
+fs.writeFileSync(path, JSON.stringify(schema, null, 2));
