+++ custom_typings/main.d.ts
@@ -4,4 +4,5 @@
 /// <reference path="./espree.d.ts" />
 /// <reference path="./estraverse.d.ts" />
 /// <reference path="./estree.d.ts" />
+/// <reference path="./jsonschema.d.ts" />
 /// <reference path="./shady-css-parser.d.ts" />
