+++ package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "clean": "mkdir -p lib/ && rm -rf lib/",
     "compile": "tsc",
-    "build": "mkdir -p lib/ && cp -r src/test lib/ && npm run compile",
+    "build": "mkdir -p lib/ && cp -r src/test lib/ && npm run compile && typson schema src/serialized-analysis.ts SerializedAnalysis > lib/analysis.schema.json && node build-tools/fixup-schema.js",
     "release": "npm run compile",
     "build:watch": "watch 'npm run build' ./src",
     "lint": "gulp lint",
@@ -50,6 +50,7 @@
     "tslint": "^3.10.2",
     "typescript": "^2.0.0",
     "typings": "^1.3.1",
+    "typson": "^0.6.0",
     "watchy": "^0.6.6"
   },
   "dependencies": {
